fix: move conver to common helpers
